+++ db/migrate/20160406180016_create_results.rb
@@ -1,14 +1,14 @@
 class CreateResults < ActiveRecord::Migration
   def change
     create_table :results do |t|
-      t.integer :math, default: 0
-      t.integer :literature, default: 0
-      t.integer :english, default: 0
-      t.integer :physical, default: 0
-      t.integer :chemistry, default: 0
-      t.integer :biological, default: 0
-      t.integer :history, default: 0
-      t.integer :geography, default: 0
+      t.float :math, default: 0
+      t.float :literature, default: 0
+      t.float :english, default: 0
+      t.float :physical, default: 0
+      t.float :chemistry, default: 0
+      t.float :biological, default: 0
+      t.float :history, default: 0
+      t.float :geography, default: 0
       t.references :examinee, index: true, foreign_key: true
       t.timestamps null: false
     end
